+++ src/main/java/org/bytedeco/javacpp/tools/DeclarationList.java
@@ -66,21 +66,23 @@ class DeclarationList extends ArrayList<Declaration> {
         boolean add = true;
         if (templateMap != null && templateMap.empty() && !decl.custom && (decl.type != null || decl.declarator != null)) {
             // method templates cannot be declared in Java, but make sure to make their
-            // info available on request (when Info.javaNames is set) to be able to create instances
+            // info available on request (when Info.javaNames or Info.define is set) to be able to create instances
+            // decl.custom is true when info has a javaText and define is false. This allows to apply a javaText
+            // to a template without instantiating it. define forces instantiation.
             if (infoIterator == null) {
                 Type type = templateMap.type = decl.type;
                 Declarator dcl = templateMap.declarator = decl.declarator;
-                for (String name : new String[] {fullName, dcl != null ? dcl.cppName : type.cppName}) {
+                for (String name : new String[] {fullName, dcl != null ? (dcl.type.constructor ? Parser.constructorName(dcl.cppName) : dcl.cppName) : type.cppName}) {
                     if (name == null) {
                         continue;
                     }
                     List<Info> infoList = infoMap.get(name);
-                    boolean hasJavaName = false;
+                    boolean hasJavaNameOrDefine = false;
                     for (Info info : infoList) {
-                        hasJavaName |= info.javaNames != null && info.javaNames.length > 0;
+                        hasJavaNameOrDefine |= info.javaNames != null && info.javaNames.length > 0 || info.define;
                     }
-                    if (!decl.function || hasJavaName) {
-                        infoIterator = infoList.size() > 0 ? infoList.listIterator() : null;
+                    if (!decl.function || hasJavaNameOrDefine) {
+                        infoIterator = infoList.size() > 0 ? new ArrayList<>(infoList).listIterator() : null;
                         break;
                     }
                 }